redo entire API
Rather than having an `algo` parameter for domain separation, fold this into
the p2c tweak hash. This is more efficient and makes that this library can
be compatible with BIP-0340 Taproot tweaks.

HOWEVER, doing so revealed a potential vector for collisions: if the same
tagged hash is used with both x-only and full keys, which are serialized
as different lengths, then a carefully crafted message can be used to
produce a x-only key and a full key which have the same tweak but different
keys/messages.

Is this a plausible attack on any real system? Hard to imagine. But to
close this off, we make sure that each hash must choose whether it supports
x-only or full keys, and the TweakHash trait does not support both.

This required rewriting pretty-much the whole library, and took a few
iterations to get right, but I'm happy with the result.

Next commit will add some Taproot tests to make sure that works, then the
next one will clean up the docs, then I think we can release.