add taproot hash and unit test
It would be nice to implement TweakHash directly for bitcoin::TapTweakHash,
but (a) that type does not implement bitcoin_hashes::Hash; (b) it would
require I add a whole rust-bitcoin dep; and (c) the version of rust-secp
and bitcoin_hashes that rust-bitcoin uses is way behind.

So instead we add our own new hash type, but do a unit test to confirm
that it produces identical output.