add serde passthrough for tweaked keys
This allows deserializing a key with an arbitrary hash type, and therefore
"casting" a key with one kind of tweak to another by going through serde.
I don't see a way to avoid this without being overly paternalistic, so we
allow it. Hopefully if users do this, they will file an issue explaining
wtf they are doing and we can offer a dedicated method with a suitable
scary name.